# fullstackwebapp/specs/002-phase-ii-frontend/contracts/api-contract.yaml
# API Contract for Phase II Frontend
# These are stub implementations as per specification

openapi: 3.0.0
info:
  title: Phase II Frontend API Stubs
  version: 1.0.0
  description: API contract for frontend-backend integration (stubs only)

paths:
  /api/signin:
    post:
      summary: Authenticate user (stub)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
      responses:
        '200':
          description: Successful authentication
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  user:
                    type: object
                    properties:
                      id:
                        type: string
                      name:
                        type: string
                      email:
                        type: string
        '400':
          description: Invalid credentials
        '401':
          description: Unauthorized

  /api/signup:
    post:
      summary: Register new user (stub)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 100
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
      responses:
        '200':
          description: User registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    type: object
                    properties:
                      id:
                        type: string
                      name:
                        type: string
                      email:
                        type: string
        '400':
          description: Invalid input data

  /api/tasks:
    get:
      summary: Get user tasks (stub)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of user tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    title:
                      type: string
                    description:
                      type: string
                    status:
                      type: string
                      enum: [pending, completed, in-progress]
                    createdDate:
                      type: string
                      format: date-time
    post:
      summary: Create new task (stub)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  minLength: 1
                  maxLength: 255
                description:
                  type: string
                  maxLength: 1000
                status:
                  type: string
                  enum: [pending, completed, in-progress]
      responses:
        '200':
          description: Task created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  title:
                    type: string
                  description:
                    type: string
                  status:
                    type: string
                  createdDate:
                    type: string
                    format: date-time

  /api/tasks/{id}:
    put:
      summary: Update task (stub)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  minLength: 1
                  maxLength: 255
                description:
                  type: string
                  maxLength: 1000
                status:
                  type: string
                  enum: [pending, completed, in-progress]
      responses:
        '200':
          description: Task updated successfully
    delete:
      summary: Delete task (stub)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Task deleted successfully

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer